resources/agents/51-local-rag-researcher-deepseek/agent/requirements.txt
langgraph
langchain-core
langchain_openai
langchain_experimental
langchain_text_splitters
langchain_huggingface
langchain_chroma
chroma
unstructured
python-dotenv
tavily-python
ollama
streamlit
streamlit_nested_layout
pdfplumber
pyperclip